Remove AltOperators.
Earlier libraries of mine offered multiple synonymous names for
operations, allowing the user to choose which was best. This, in
the end, seemed rather cluttered and suboptimal. I find the |+|
notation easier to read (| is bigger than .) and to predict, so
I'd like to keep just that. A user really clamoring for different
syntax can always write it themselves.